Phoning Home
Progress - Sort of. Thanks to Michael, there have been huge leaps
forward in how the widget could get it’s data - it’s just not really
working yet. Enter username and password, then run the widget in Safari
to see the log files in the inspector - There is clearly a lot of data
there. Now it just remains to put it in the table. Somehow.